Take snapshot for dark mode in Xcode 10 with iPhone Xr
I figured out that it was the iPhone Xr the one to take the reference
snapshots, but then the test pass while using Xcode 10, but not for
Xcode 11.

The iOS 13 SDK seems to change some margins.

The reference image for dark mode shows as light mode because iOS 12
doesn't support dark mode yet, then by adding this new reference the
test will pass until the library is moved to Xcode 11. Then you can
update the reference image.