Fix crash when using `UIWindow` in `assertSnapshot`
`prepareView` in `View.swift` will now check if the current snapshot
target is a `UIWindow`. If that condition is `true`, then it will return
early and use the window as its dispose value. It will make the assumption
that the window has a `rootViewController` set and use that as it's `viewController`.

This avoids unbalanced calls to appearance transitions during test teardown
which will throw exceptions and crash the test.