Support full customization of artifacts directory
This is especially useful when using swift-snapshot-testing with an Objective-C wrapper. In this case, the `file: StaticString` is not the path to the file where the test is but rather the wrapper. Allowing full customization allows us to pass an appropriate artifacts directory based on the Objective-C test case name.