`assertSnapshot`: added separate parameter to override snapshot file
Currently the same `#file` parameter is used both to determine the name of the snapshot file, as well as the parameter passed to `XCTFail`.

However for certain workflows, these might be 2 different things.
For example, [RevenueCat](https://github.com/RevenueCat/purchases-ios/blob/4.10.0/Tests/UnitTests/Mocks/MockHTTPClient.swift#L66-L69) overrides this to dynamically set the file to whatever test is running, because `assertSnapshot` is called from a separate file. However, when the assertion fails, the error also comes from that file, which has an erroneous `#line`. Instead, this would allow passing `testFile` only to override the name of the snapshot file, without changing the `file`/`line` where the failure would come.