Write to snapshot state before failing
XCTest test cases can be configured with `continueAfterFailure = false`,
and in fact UI tests are configured with this by default, but in these
cases inline snapshots will never be written because the failure aborts
the test before we write to the snapshot state.

This commit reorders things to fix the problem.