Xcode 16 beta 5: Fix snapshots test trait
It is no longer possible to `@_spi(Experimental) import Testing`, so we
can no longer ship a custom execution test trait.

Instead of using such a trait to override a task local for the duration
of a test, we can use a more general trait to store the information, and
the assertion helpers can coalesce to the trait, instead.