Introduce SnapshotUITesting module for UITest support
Swift Testing is not compatible with UITest targets at the moment, but
unfortunately `#if canImport` still returns `true`, making it difficult
to ship libraries that provide tools for both `Testing` and `XCTest`,
like the `assertSnapshot`. The previous workarounds we relied on have
stopped working, and so we need a new approach.

In this PR we have introduced a forwards-compatible `SnapshotUITesting`
module that does not rely on any imports of `Testing` internally. We can
then introduce Swift Testing support in the `SnapshotTesting` module,
instead.